Move prepaid agreements code into its own app
