+++ cfgov/agreements/urls.py
@@ -1,11 +1,9 @@
 from django.conf.urls import url
 
-from agreements.views import index, issuer_search, prepaid, detail
+from agreements.views import index, issuer_search
 
 
 urlpatterns = [
-    url(r'^$', prepaid, name='prepaid'),
-    url(r'^detail/(?P<product_id>\d+)/$', detail, name='detail'),
     url(r'^$', index, name='agreements_home'),
     url(r'^issuer/(?P<issuer_slug>.*)/$',
         issuer_search,
