+++ cfgov/prepaid_agreements/jinja2/prepaid_agreements/index.html
@@ -3,7 +3,7 @@
 {% import 'atoms/radio-button.html' as radio %}
 {% import 'atoms/tag.html' as tag %}
 <!-- {% import 'regulations3k/regulations3k-search-bar.html' as search_bar %} -->
-{% import 'agreements/search_result_item.html' as search_item %}
+{% import 'prepaid_agreements/search_result_item.html' as search_item %}
 {% import 'molecules/pagination.html' as pagination with context %}
 {% from 'organisms/expandable.html' import expandable with context %}
